feat: entities, repos and resources; models to entities; controllers to resources; long id

+++ src/main/java/br/ufes/sead/sgcbackend/repositories/AddressRepository.java
@@ -3,9 +3,8 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Address;
+import br.ufes.sead.sgcbackend.entities.Address;
 
-public interface AddressRepository
-        extends CrudRepository<Address, Integer>, PagingAndSortingRepository<Address, Integer> {
+public interface AddressRepository extends CrudRepository<Address, Long>, PagingAndSortingRepository<Address, Long> {
 
 }
