+++ src/main/java/br/ufes/sead/sgcbackend/repositories/ApplicantRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Applicant;
+import br.ufes.sead.sgcbackend.entities.Applicant;
 
 public interface ApplicantRepository
-        extends CrudRepository<Applicant, Integer>, PagingAndSortingRepository<Applicant, Integer> {
+                extends CrudRepository<Applicant, Long>, PagingAndSortingRepository<Applicant, Long> {
 
 }
