+++ src/main/java/br/ufes/sead/sgcbackend/repositories/BondRepository.java
@@ -3,8 +3,8 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Bond;
+import br.ufes.sead.sgcbackend.entities.Bond;
 
-public interface BondRepository extends CrudRepository<Bond, Integer>, PagingAndSortingRepository<Bond, Integer> {
+public interface BondRepository extends CrudRepository<Bond, Long>, PagingAndSortingRepository<Bond, Long> {
 
 }
