+++ src/main/java/br/ufes/sead/sgcbackend/repositories/CourseClassRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.CourseClass;
+import br.ufes.sead.sgcbackend.entities.CourseClass;
 
 public interface CourseClassRepository
-        extends CrudRepository<CourseClass, Integer>, PagingAndSortingRepository<CourseClass, Integer> {
+                extends CrudRepository<CourseClass, Long>, PagingAndSortingRepository<CourseClass, Long> {
 
 }
