+++ src/main/java/br/ufes/sead/sgcbackend/repositories/CourseRepository.java
@@ -3,8 +3,8 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Course;
+import br.ufes.sead.sgcbackend.entities.Course;
 
-public interface CourseRepository extends CrudRepository<Course, Integer>, PagingAndSortingRepository<Course, Integer> {
+public interface CourseRepository extends CrudRepository<Course, Long>, PagingAndSortingRepository<Course, Long> {
 
 }
