+++ src/main/java/br/ufes/sead/sgcbackend/repositories/DocumentRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Document;
+import br.ufes.sead.sgcbackend.entities.Document;
 
 public interface DocumentRepository
-        extends CrudRepository<Document, Integer>, PagingAndSortingRepository<Document, Integer> {
+                extends CrudRepository<Document, Long>, PagingAndSortingRepository<Document, Long> {
 
 }
