+++ src/main/java/br/ufes/sead/sgcbackend/repositories/DocumentTypeRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.DocumentType;
+import br.ufes.sead.sgcbackend.entities.DocumentType;
 
 public interface DocumentTypeRepository
-        extends CrudRepository<DocumentType, Integer>, PagingAndSortingRepository<DocumentType, Integer> {
+                extends CrudRepository<DocumentType, Long>, PagingAndSortingRepository<DocumentType, Long> {
 
 }
