+++ src/main/java/br/ufes/sead/sgcbackend/repositories/EmployeeRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Employee;
+import br.ufes.sead.sgcbackend.entities.Employee;
 
 public interface EmployeeRepository
-        extends CrudRepository<Employee, Integer>, PagingAndSortingRepository<Employee, Integer> {
+                extends CrudRepository<Employee, Long>, PagingAndSortingRepository<Employee, Long> {
 
 }
