+++ src/main/java/br/ufes/sead/sgcbackend/repositories/IdentityRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Identity;
+import br.ufes.sead.sgcbackend.entities.Identity;
 
 public interface IdentityRepository
-        extends CrudRepository<Identity, Integer>, PagingAndSortingRepository<Identity, Integer> {
+                extends CrudRepository<Identity, Long>, PagingAndSortingRepository<Identity, Long> {
 
 }
