+++ src/main/java/br/ufes/sead/sgcbackend/repositories/ImpedimentRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Impediment;
+import br.ufes.sead.sgcbackend.entities.Impediment;
 
 public interface ImpedimentRepository
-        extends CrudRepository<Impediment, Integer>, PagingAndSortingRepository<Impediment, Integer> {
+                extends CrudRepository<Impediment, Long>, PagingAndSortingRepository<Impediment, Long> {
 
 }
