+++ src/main/java/br/ufes/sead/sgcbackend/repositories/PhoneRepository.java
@@ -3,8 +3,8 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Phone;
+import br.ufes.sead.sgcbackend.entities.Phone;
 
-public interface PhoneRepository extends CrudRepository<Phone, Integer>, PagingAndSortingRepository<Phone, Integer> {
+public interface PhoneRepository extends CrudRepository<Phone, Long>, PagingAndSortingRepository<Phone, Long> {
 
 }
