+++ src/main/java/br/ufes/sead/sgcbackend/repositories/RoleRepository.java
@@ -3,8 +3,8 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.Role;
+import br.ufes.sead.sgcbackend.entities.Role;
 
-public interface RoleRepository extends CrudRepository<Role, Integer>, PagingAndSortingRepository<Role, Integer> {
+public interface RoleRepository extends CrudRepository<Role, Long>, PagingAndSortingRepository<Role, Long> {
 
 }
