+++ src/main/java/br/ufes/sead/sgcbackend/repositories/UserRepository.java
@@ -3,8 +3,8 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.User;
+import br.ufes.sead.sgcbackend.entities.User;
 
-public interface UserRepository extends CrudRepository<User, Integer>, PagingAndSortingRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long>, PagingAndSortingRepository<User, Long> {
 
 }
