+++ src/main/java/br/ufes/sead/sgcbackend/repositories/UserTypeRepository.java
@@ -3,9 +3,9 @@ package br.ufes.sead.sgcbackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import br.ufes.sead.sgcbackend.models.UserType;
+import br.ufes.sead.sgcbackend.entities.UserType;
 
 public interface UserTypeRepository
-        extends CrudRepository<UserType, Integer>, PagingAndSortingRepository<UserType, Integer> {
+                extends CrudRepository<UserType, Long>, PagingAndSortingRepository<UserType, Long> {
 
 }
